idcs_domain/infaas/inference/pyproject.toml
[project]
name = "inference"
version = "0.1.0"
description = ""
authors = [
    {name = "GalSarid21",email = "gal.sarid@intel.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.12"
dependencies = [
    "aiohttp (>=3.11.11,<4.0.0)",
    "aiosignal (>=1.3.2,<2.0.0)",
    "annotated-types (>=0.7.0,<0.8.0)",
    "async-timeout (>=5.0.1,<6.0.0)",
    "attrs (>=25.1.0,<26.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)",
    "filelock (>=3.17.0,<4.0.0)",
    "frozenlist (>=1.5.0,<2.0.0)",
    "fsspec (>=2025.2.0,<2026.0.0)",
    "httpx (==0.27.2)",
    "huggingface-hub (>=0.28.1,<0.29.0)",
    "idna (>=3.10,<4.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "multidict (>=6.1.0,<7.0.0)",
    "openai (>=1.61.0,<2.0.0)",
    "packaging (>=24.2,<25.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-core (>=2.28.0,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "safetensors (>=0.5.2,<0.6.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "setuptools (>=75.8.0,<76.0.0)",
    "text-generation (>=0.7.0,<0.8.0)",
    "tokenizers (>=0.21.0,<0.22.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "transformers (>=4.48.2,<5.0.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "validators (>=0.34.0,<0.35.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "yarl (>=1.18.3,<2.0.0)",
    "grpcio (==1.65.5)",
    "grpcio-health-checking (==1.65.5)",
    "grpcio-reflection (==1.65.1)",
    "grpcio-tools (==1.65.1)",
    "protobuf (==5.27.2)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "."


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
nltk = "^3.9.1"
locust = "^2.32.8"
pytest-asyncio = "^0.25.3"
numpy = "^2.2.2"
huggingface-hub = "^0.28.1"
torch = "2.5.0"
boto3 = "^1.36.17"
tqdm = "^4.67.1"


[tool.poetry]
package-mode = false
